Dossier_Streamlit/requirements.txt
pandas
seaborn
streamlit
bs4
requestsc